vector-proto/requirements.txt
PyPDF2
langchain
openai
faiss-cpu
tiktoken
streamlit-authenticator